--- pandas/tests/series/test_analytics.py
@@ -1,23 +1,0 @@
-import numpy as np
-
-import pandas as pd
-from pandas import Series
-
-
-class TestSeriesAnalytics:
-    def test_is_monotonic(self):
-
-        s = Series(np.random.randint(0, 10, size=1000))
-        assert not s.is_monotonic
-        s = Series(np.arange(1000))
-        assert s.is_monotonic is True
-        assert s.is_monotonic_increasing is True
-        s = Series(np.arange(1000, 0, -1))
-        assert s.is_monotonic_decreasing is True
-
-        s = Series(pd.date_range("20130101", periods=10))
-        assert s.is_monotonic is True
-        assert s.is_monotonic_increasing is True
-        s = Series(list(reversed(s.tolist())))
-        assert s.is_monotonic is False
-        assert s.is_monotonic_decreasing is True
