+++ pandas/tests/series/test_arithmetic.py
@@ -15,6 +15,7 @@ from pandas import (
     Index,
     IntervalIndex,
     Series,
+    Timedelta,
     bdate_range,
     date_range,
     isna,
@@ -276,6 +277,25 @@ class TestSeriesArithmetic:
 
         assert ser.index is dti
         assert ser_utc.index is dti_utc
+
+    def test_arithmetic_with_duplicate_index(self):
+
+        # GH#8363
+        # integer ops with a non-unique index
+        index = [2, 2, 3, 3, 4]
+        ser = Series(np.arange(1, 6, dtype="int64"), index=index)
+        other = Series(np.arange(5, dtype="int64"), index=index)
+        result = ser - other
+        expected = Series(1, index=[2, 2, 3, 3, 4])
+        tm.assert_series_equal(result, expected)
+
+        # GH#8363
+        # datetime ops with a non-unique index
+        ser = Series(date_range("20130101 09:00:00", periods=5), index=index)
+        other = Series(date_range("20130101", periods=5), index=index)
+        result = ser - other
+        expected = Series(Timedelta("9 hours"), index=[2, 2, 3, 3, 4])
+        tm.assert_series_equal(result, expected)
 
 
 # ------------------------------------------------------------------
